feat(process): do not start stopped processes
If a process is started and stopped within the same stack frame, do not
begin the process. This change ensures that the first step(s) in a
process do not run after the process has been stopped; otherwise, any
cleanup tasks scheduled after the process stops may run BEFORE an action
has started. In any case, no actions should run after a process is
stopped. This is a breaking change because it modifies the timing and
order of process actions.

BREAKING