fix(data): pass request to proxy.url
When no proxy rules exist to provide a protocol or port but the Request
contains a protocol or port provided by the original DataDefinition
object, those values were being ignored and defaulted. Modifying the
Proxy#url function to accept an entire Request objects as an alternative
to just the base and paths enables it to construct a final URL that
respects the protocol and port specified in the original definition.