feat(data): change proxy origin behaviors
The proxy will no longer use the data definition's `base` value as the
host name when no rules exist to set the request's origin. Instead, the
request will be treated as relative (i.e. the URL will start with a /
and contain the full path, if specified).

Additionally, if the developer wishes to specify their origin as a
single string (protocol + host + port) rather than as individual
properties, they can now do so through a new `origin` property on
ProxyRule.

BREAKING CHANGE