fix: allow more lenient data definition schema
After we transitioned to using relative paths for proxy rules, having a
data definition object without a base was possible. But also, it's
always possible that someone wants to POST or GET from an endpoint that
doesn't contain a path portion. So now we just check if the definition
is object-like.